Replace webpack dist build with rollup
