+++ docs/index.js.html
@@ -37,6 +37,13 @@
 				</ul>
 			</li>
 			
+			<li class="dropdown">
+				<a href="global.html" class="dropdown-toggle" data-toggle="dropdown">Global<b class="caret"></b></a>
+				<ul class="dropdown-menu inline">
+					<li><a href="global.html#getArrayKey">getArrayKey</a></li>
+				</ul>
+			</li>
+			
 		</ul>
         
             <div class="col-sm-3 col-md-3">
@@ -81,17 +88,14 @@ import {DEFAULT_OPTIONS} from './constants';
 // instance
 import {augmentMoizeInstance} from './instance';
 
-// is equal
-import {getIsEqual} from './isEqual';
-
 // max age
 import {getMaxAgeOptions} from './maxAge';
 
+// is equal
+import {getIsEqual, getTransformKey} from './options';
+
 // stats
 import {collectStats, getDefaultProfileName, getStats, getStatsOptions, statsCache} from './stats';
-
-// transform key
-import {getTransformKey} from './transformKey';
 
 // types
 import type {Expiration, MicroMemoizeOptions, Options} from './types';
