+++ package.json
@@ -3,32 +3,16 @@
   "ava": {
     "babel": {
       "babelrc": true,
-      "presets": [
-        "react"
-      ]
+      "presets": ["react"]
     },
     "failFast": true,
-    "files": [
-      "test/*.js",
-      "!test-utils/*.js"
-    ],
-    "require": [
-      "babel-register",
-      "babel-polyfill"
-    ],
-    "source": [
-      "src/*.js"
-    ],
+    "files": ["test/*.js", "!test-utils/*.js"],
+    "require": ["babel-register", "babel-polyfill"],
+    "source": ["src/*.js"],
     "verbose": true
   },
-  "browserslist": [
-    "defaults",
-    "Explorer >= 9",
-    "Safari >= 6",
-    "Opera >= 15",
-    "iOS >= 8",
-    "Android >= 4"
-  ],
+  "browser": "dist/moize.min.js",
+  "browserslist": ["defaults", "Explorer >= 9", "Safari >= 6", "Opera >= 15", "iOS >= 8", "Android >= 4"],
   "bugs": {
     "url": "https://github.com/planttheidea/moize/issues"
   },
@@ -40,7 +24,6 @@
     "babel-loader": "^7.1.4",
     "babel-plugin-add-module-exports": "0.2.1",
     "babel-plugin-syntax-flow": "^6.18.0",
-    "babel-plugin-transform-decorators-legacy": "1.3.4",
     "babel-plugin-transform-flow-strip-types": "6.22.0",
     "babel-polyfill": "^6.26.0",
     "babel-preset-env": "1.6.1",
@@ -79,11 +62,16 @@
     "react": "16.2.0",
     "react-dom": "16.2.0",
     "rimraf": "2.6.2",
+    "rollup": "0.56.5",
+    "rollup-plugin-babel": "3.0.3",
+    "rollup-plugin-commonjs": "9.0.0",
+    "rollup-plugin-node-resolve": "3.2.0",
+    "rollup-plugin-uglify": "3.0.0",
     "sinon": "^4.4.2",
     "underscore": "1.8.3",
     "webpack": "^4.1.1",
-    "webpack-cli": "2.0.10",
-    "webpack-dev-server": "^3.1.0"
+    "webpack-cli": "2.0.11",
+    "webpack-dev-server": "^3.1.1"
   },
   "homepage": "https://github.com/planttheidea/moize#readme",
   "keywords": [
@@ -116,23 +104,23 @@
   },
   "scripts": {
     "benchmark": "npm run transpile:lib -- --no-comments && node benchmark/index.js",
-    "build": "NODE_ENV=development webpack --progress --colors",
-    "build:minified": "NODE_ENV=production webpack --progress --colors --config=webpack.config.minified.js",
-    "build:stats": "NODE_ENV=production webpack --progress --colors --config=webpack.config.minified.js --json > stats.json",
+    "build": "NODE_ENV=production rollup -c",
     "clean": "npm run clean:lib && npm run clean:dist && npm run clean:docs",
     "clean:docs": "rimraf docs",
     "clean:dist": "rimraf dist",
     "clean:lib": "rimraf lib",
     "clean:es": "rimraf es",
-    "dev": "NODE_ENV=development webpack-dev-server --colors --progress --config=webpack.config.dev.js",
-    "dist": "npm run clean:dist && npm run build && npm run build:minified",
+    "dev": "NODE_ENV=development webpack-dev-server --colors --progress --config=webpack/webpack.config.js",
+    "dist": "npm run clean:dist && npm run build",
     "docs": "npm run clean:docs && jsdoc -c ./jsdoc.config.json && cp img/*.png docs/img && git add --all",
     "flow": "flow check src",
     "flow:generate": "flowgen ./index.d.ts --output-file flow-typed/moize.js",
     "lint": "NODE_ENV=test eslint src",
     "lint:fix": "NODE_ENV=test eslint src --fix",
+    "postpublish": "npm run docs",
     "prepublish": "if in-publish; then npm run prepublish:compile; fi",
-    "prepublish:compile": "npm run lint && npm run flow && npm run test:coverage && npm run transpile:lib && npm run transpile:es && npm run dist && npm run docs",
+    "prepublish:compile":
+      "npm run lint && npm run flow && npm run test:coverage && npm run transpile:lib && npm run transpile:es && npm run dist",
     "start": "npm run dev",
     "test": "NODE_PATH=. NODE_ENV=production BABEL_ENV=test ava",
     "test:coverage": "nyc npm test",
